ATD Advanced Stats Bot/requirements.txt
discord.py>=2.3.0
nba-api>=1.5.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0